[arcane,c#tools] Déplace les classes 'CommandArgs' et 'ExecDriverProperties' dans leur propre fichier.
